+++ crates/web-client/test/fpi.test.ts
@@ -31,12 +31,10 @@ export const testStandardFpi = async (page: Page): Promise<void> => {
                 swapw dropw
             end
         `;
-
-    let getItemComponent = window.AccountComponent.compile(
-      code,
-      window.TransactionKernel.assembler(),
-      [window.StorageSlot.map(storageMap)]
-    ).withSupportsAllTypes();
+    let builder = client.createScriptBuilder();
+    let getItemComponent = window.AccountComponent.compile(code, builder, [
+      window.StorageSlot.map(storageMap),
+    ]).withSupportsAllTypes();
 
     const walletSeed = new Uint8Array(32);
     crypto.getRandomValues(walletSeed);
@@ -107,10 +105,7 @@ export const testStandardFpi = async (page: Page): Promise<void> => {
         foreignAccountId.prefix().asInt().toString()
       );
 
-    let compiledTxScript = window.TransactionScript.compile(
-      txScript,
-      window.TransactionKernel.assembler()
-    );
+    let compiledTxScript = builder.compileTxScript(txScript);
 
     await client.syncState();