+++ crates/web-client/test/new_transactions.test.ts
@@ -347,7 +347,6 @@ export const customTransaction = async (
 ): Promise<void> => {
   return await testingPage.evaluate(
     async ({ assertedValue, withRemoteProver }) => {
-      debugger;
       const client = window.client;
 
       const walletAccount = await client.newWallet(
@@ -472,7 +471,8 @@ export const customTransaction = async (
             end
         `;
 
-      let compiledNoteScript = await client.compileNoteScript(noteScript);
+      let builder = client.createScriptBuilder();
+      let compiledNoteScript = builder.compileNoteScript(noteScript);
       let noteInputs = new window.NoteInputs(
         new window.FeltArray([
           walletAccount.id().prefix(),
@@ -532,7 +532,7 @@ export const customTransaction = async (
 
       // Creating Second Custom Transaction Request to Consume Custom Note
       // with Invalid/Valid Transaction Script
-      let transactionScript = await client.compileTxScript(txScript);
+      let transactionScript = await builder.compileTxScript(txScript);
       let noteArgsCommitment = window.Rpo256.hashElements(feltArray); // gets consumed by NoteIdAndArgs
       let noteAndArgs = new window.NoteAndArgs(note, noteArgsCommitment);
       let noteAndArgsArray = new window.NoteAndArgsArray([noteAndArgs]);
@@ -787,14 +787,14 @@ export const customAccountComponent = async (
         end
       `;
     const client = window.client;
-    let assembler = window.TransactionKernel.assembler().withDebugMode(true);
+    let builder = client.createScriptBuilder();
     let emptyStorageSlot = window.StorageSlot.emptyValue();
     let storageMap = new window.StorageMap();
     let storageSlotMap = window.StorageSlot.map(storageMap);
 
     let mappingAccountComponent = window.AccountComponent.compile(
       accountCode,
-      assembler,
+      builder,
       [emptyStorageSlot, storageSlotMap]
     ).withSupportsAllTypes();
 
@@ -816,17 +816,14 @@ export const customAccountComponent = async (
 
     await client.syncState();
 
-    let accountComponentLib =
-      window.AssemblerUtils.createAccountComponentLibrary(
-        assembler,
-        "miden_by_example::mapping_example_contract",
-        accountCode
-      );
-
-    let txScript = window.TransactionScript.compile(
-      scriptCode,
-      assembler.withLibrary(accountComponentLib)
+    let accountCodeLib = builder.buildLibrary(
+      "miden_by_example::mapping_example_contract",
+      accountCode
     );
+
+    builder.linkStaticLibrary(accountCodeLib);
+
+    let txScript = builder.compileTxScript(scriptCode);
 
     let txIncrementRequest = new window.TransactionRequestBuilder()
       .withCustomScript(txScript)
@@ -1093,12 +1090,13 @@ export const counterAccountComponent = async (
     const client = window.client;
 
     // Create counter account
-    let assembler = window.TransactionKernel.assembler().withDebugMode(true);
     let emptyStorageSlot = window.StorageSlot.emptyValue();
+
+    let builder = client.createScriptBuilder();
 
     let counterAccountComponent = window.AccountComponent.compile(
       accountCode,
-      assembler,
+      builder,
       [emptyStorageSlot]
     ).withSupportsAllTypes();
 
@@ -1121,17 +1119,12 @@ export const counterAccountComponent = async (
     await client.syncState();
 
     // Deploy counter account
-    let accountComponentLib =
-      window.AssemblerUtils.createAccountComponentLibrary(
-        assembler,
-        "external_contract::counter_contract",
-        accountCode
-      );
-
-    let txScript = window.TransactionScript.compile(
-      scriptCode,
-      assembler.withLibrary(accountComponentLib)
+    let accountComponentLib = builder.buildLibrary(
+      "external_contract::counter_contract",
+      accountCode
     );
+    builder.linkDynamicLibrary(accountComponentLib);
+    let txScript = builder.compileTxScript(scriptCode);
 
     let txIncrementRequest = new window.TransactionRequestBuilder()
       .withCustomScript(txScript)
@@ -1147,11 +1140,7 @@ export const counterAccountComponent = async (
     );
 
     // Create transaction with network note
-    assembler = window.TransactionKernel.assembler()
-      .withDebugMode(true)
-      .withLibrary(accountComponentLib);
-
-    let compiledNoteScript = await assembler.compileNoteScript(scriptCode);
+    let compiledNoteScript = await builder.compileNoteScript(scriptCode);
 
     let noteInputs = new window.NoteInputs(new window.FeltArray([]));
 
@@ -1207,6 +1196,7 @@ test.describe("counter account component tests", () => {
   test("counter account component transaction completes successfully", async ({
     page,
   }) => {
+    page.on("console", (msg) => console.log(msg));
     let finalCounter = await counterAccountComponent(page);
     expect(finalCounter).toEqual("2");
   });
@@ -1250,9 +1240,10 @@ export const testStorageMap = async (page: Page): Promise<any> => {
                 end
         `;
 
+    let builder = client.createScriptBuilder();
     let bumpItemComponent = window.AccountComponent.compile(
       accountCode,
-      window.TransactionKernel.assembler(),
+      builder,
       [window.StorageSlot.map(storageMap)]
     ).withSupportsAllTypes();
 
@@ -1279,21 +1270,19 @@ export const testStorageMap = async (page: Page): Promise<any> => {
       ?.toHex();
 
     // Deploy counter account
-    let assembler = window.TransactionKernel.assembler();
 
-    let accountComponentLib =
-      window.AssemblerUtils.createAccountComponentLibrary(
-        assembler,
-        "external_contract::bump_item_contract",
-        accountCode
-      );
+    let accountComponentLib = builder.buildLibrary(
+      "external_contract::bump_item_contract",
+      accountCode
+    );
 
-    let txScript = window.TransactionScript.compile(
+    builder.linkDynamicLibrary(accountComponentLib);
+
+    let txScript = builder.compileTxScript(
       `use.external_contract::bump_item_contract
       begin
           call.bump_item_contract::bump_map_item
-      end`,
-      assembler.withLibrary(accountComponentLib)
+      end`
     );
 
     let txIncrementRequest = new window.TransactionRequestBuilder()